requirements.txt
# Core Dependencies
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0

# Optional (for enhanced functionality)
pandas==2.1.3  # For data analysis if needed
gunicorn==21.2.0  # For production deployment
